+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Toggle all button to Checkbox and Select (multi only) element
+
 ### Changed
 
 - Barcode Scanner lookups to run automatically on scan